refactor(amp): move every amp handler to separate file
