+++ amp/handlers/index.js
@@ -1,0 +1,4 @@
+const { ampStoryPageHandler } = require("./story-page");
+const { storyPageInfiniteScrollHandler } = require("./infinite-scroll");
+
+module.exports = { ampStoryPageHandler, storyPageInfiniteScrollHandler };
