+++ amp/helpers/index.js
@@ -2,4 +2,13 @@ const { InfiniteScrollAmp } = require("./infinite-scroll");
 const { setCorsHeaders } = require("./set-cors-headers");
 const { optimize } = require("./optimize-amp-html");
 
-module.exports = { InfiniteScrollAmp, setCorsHeaders, optimize };
+function getSeoInstance(seo, config, pageType = "") {
+  return typeof seo === "function" ? seo(config, pageType) : seo;
+}
+
+module.exports = {
+  InfiniteScrollAmp,
+  setCorsHeaders,
+  optimize,
+  getSeoInstance,
+};
